* add factory method for Mapper maps * move ClusterConfigSpecExMap to separate file
